Revert "feat: Admin CRUD 작성 및 테스트 완료, Notice CRUD 작성 및 테스트 완료"
This reverts commit 277aef70195a0b1b29a5a3a7613a976ddd5ce9cf.

+++ src/main/java/com/jackpot/mapper/NoticeMapper.java
@@ -3,11 +3,11 @@ package com.jackpot.mapper;
 import com.jackpot.domain.NoticeVO;
 
 public interface NoticeMapper {
-	public NoticeVO read(Long noticeId); // PK 전달
+	public NoticeVO read(int noticeId); // PK 전달
 	
 	public void insert(NoticeVO notice);
 	
 	public int update(NoticeVO notice);
 	
-	public int delete(Long noticeId); // PK 전달
+	public int delete(int noticeId); // PK 전달
 }
